Moved PSA serializers in the PSA app and editing serverside for the Corvet table
