+++ psa/templates/psa/corvet_table.html
@@ -35,6 +35,9 @@
 {% block tablejs %}
 
     <!-- Page level custom scripts -->
+    <script type="text/javascript">
+        const URL_AJAX = '/psa/api/corvet/';
+    </script>
     <script type="text/javascript" src="{% static 'psa/js/corvet-tables.js' %}"></script>
 
     <!-- SearchPanes scripts -->
