+++ psa/urls.py
@@ -1,10 +1,14 @@
-from django.urls import path
-
+from django.urls import path, include
+from rest_framework.routers import DefaultRouter
 from . import views
+
+router = DefaultRouter()
+router.register(r'api/corvet', views.CorvetViewSet, basename='api_corvet')
 
 app_name = 'psa'
 
 urlpatterns = [
+    path('', include(router.urls)),
     path('nac/tools/', views.nac_tools, name='nac_tools'),
     path('nac/tools/license/', views.nac_license, name='nac_license'),
     path('nac/tools/update/', views.nac_update, name='nac_update'),
